Move `floki_root` resolution into `Environment` structure
This helps reduce the amount of parameter passing that is needed in functions
which figure out how to launch the container. The location of the configuration
file, and the directory it is located in (or which needs to be mounted into
the container) is part of how `floki` is run, so belongs in the `Environment`
structure.

This commit doesn't fundamentally change the function in any way, it just moves
it.
