+++ src/environment.rs
@@ -1,3 +1,4 @@
+use crate::errors;
 /// Query the current user environment
 use failure::Error;
 use std::env;
@@ -8,15 +9,20 @@ use std::process::Command;
 pub struct Environment {
     pub user_details: (String, String),
     pub current_directory: path::PathBuf,
+    pub floki_root: path::PathBuf,
+    pub config_file: path::PathBuf,
     pub ssh_agent_socket: Option<String>,
 }
 
 impl Environment {
     /// Gather information on the environment floki is running in
-    pub fn gather() -> Result<Self, Error> {
+    pub fn gather(config_file: &Option<path::PathBuf>) -> Result<Self, Error> {
+        let (floki_root, config_path) = resolve_floki_root_and_config(config_file)?;
         Ok(Environment {
             user_details: get_user_details()?,
             current_directory: get_current_working_directory()?,
+            floki_root: floki_root,
+            config_file: config_path,
             ssh_agent_socket: get_ssh_agent_socket_path(),
         })
     }
@@ -48,5 +54,84 @@ fn get_ssh_agent_socket_path() -> Option<String> {
     match env::var("SSH_AUTH_SOCK") {
         Ok(path) => Some(path),
         Err(_) => None,
+    }
+}
+
+/// Search all ancestors of the current directory for a floki.yaml file name.
+fn find_floki_yaml(current_directory: &path::Path) -> Result<path::PathBuf, Error> {
+    current_directory
+        .ancestors()
+        .map(|a| a.join("floki.yaml"))
+        .find(|f| f.is_file())
+        .ok_or(errors::FlokiError::ProblemFindingConfigYaml {}.into())
+}
+
+/// Take a file path, and return a tuple consisting of it's parent directory and the file path
+fn locate_file_in_parents(path: path::PathBuf) -> Result<(path::PathBuf, path::PathBuf), Error> {
+    let dir = path
+        .parent()
+        .ok_or_else(|| errors::FlokiInternalError::InternalAssertionFailed {
+            description: format!("config_file '{:?}' does not have a parent", &path),
+        })?
+        .to_path_buf();
+    Ok((dir, path))
+}
+
+/// Resolve floki root directory and path to configuration file
+fn resolve_floki_root_and_config(
+    config_file: &Option<path::PathBuf>,
+) -> Result<(path::PathBuf, path::PathBuf), Error> {
+    match config_file {
+        Some(path) => Ok((get_current_working_directory()?, path.clone())),
+        None => Ok(locate_file_in_parents(find_floki_yaml(
+            &get_current_working_directory()?,
+        )?)?),
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use failure::format_err;
+    use std::fs;
+    use tempdir;
+
+    fn touch_file(path: &path::Path) -> Result<(), Error> {
+        fs::create_dir_all(
+            path.parent()
+                .ok_or(format_err!("Unable to take parent of path"))?,
+        )?;
+        fs::OpenOptions::new().create(true).write(true).open(path)?;
+        Ok(())
+    }
+
+    #[test]
+    fn test_find_floki_yaml_current_dir() -> Result<(), Error> {
+        let tmp_dir = tempdir::TempDir::new("")?;
+        let floki_yaml_path = tmp_dir.path().join("floki.yaml");
+        touch_file(&floki_yaml_path)?;
+        assert_eq!(find_floki_yaml(&tmp_dir.path())?, floki_yaml_path);
+        Ok(())
+    }
+
+    #[test]
+    fn test_find_floki_yaml_ancestor() -> Result<(), Error> {
+        let tmp_dir = tempdir::TempDir::new("")?;
+        let floki_yaml_path = tmp_dir.path().join("floki.yaml");
+        touch_file(&floki_yaml_path)?;
+        assert_eq!(
+            find_floki_yaml(&tmp_dir.path().join("dir/subdir"))?,
+            floki_yaml_path
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn test_find_floki_yaml_sibling() -> Result<(), Error> {
+        let tmp_dir = tempdir::TempDir::new("")?;
+        let floki_yaml_path = tmp_dir.path().join("src/floki.yaml");
+        touch_file(&floki_yaml_path)?;
+        assert!(find_floki_yaml(&tmp_dir.path().join("include")).is_err());
+        Ok(())
     }
 }
