+++ src/main.rs
@@ -20,7 +20,6 @@ use verify::verify_command;
 
 use failure::Error;
 use quicli::prelude::*;
-use std::path;
 use structopt::StructOpt;
 
 fn main() -> CliResult {
@@ -41,16 +40,12 @@ fn main() -> CliResult {
 fn run_floki_from_args(args: &Cli) -> Result<(), Error> {
     debug!("Got command line arguments: {:?}", &args);
 
-    let environ = environment::Environment::gather()?;
+    let environ = environment::Environment::gather(&args.config_file)?;
     debug!("Got environment {:?}", &environ);
 
-    let (floki_root, config_file) = match &args.config_file {
-        Some(config_file) => (environ.current_directory.clone(), config_file.clone()),
-        None => locate_file_in_system(find_floki_yaml(&environ.current_directory)?)?,
-    };
-    debug!("Selected configuration file: {:?}", &config_file);
+    debug!("Selected configuration file: {:?}", &environ.config_file);
 
-    let config = FlokiConfig::from_file(&config_file)?;
+    let config = FlokiConfig::from_file(&environ.config_file)?;
     verify_command(args.local, &config)?;
 
     // Dispatch appropriate subcommand
@@ -65,13 +60,13 @@ fn run_floki_from_args(args: &Cli) -> Result<(), Error> {
         // Run a command in the floki container
         Some(Subcommand::Run { command }) => {
             let inner_command = interpret::command_in_shell(config.shell.inner_shell(), &command);
-            run_floki_container(&environ, &floki_root, &config, inner_command)
+            run_floki_container(&environ, &config, inner_command)
         }
 
         // Launch an interactive floki shell (the default)
         None => {
             let inner_command = config.shell.inner_shell().to_string();
-            run_floki_container(&environ, &floki_root, &config, inner_command)
+            run_floki_container(&environ, &config, inner_command)
         }
     }
 }
@@ -79,79 +74,11 @@ fn run_floki_from_args(args: &Cli) -> Result<(), Error> {
 /// Launch a floki container running the inner command
 fn run_floki_container(
     environ: &environment::Environment,
-    floki_root: &path::Path,
     config: &FlokiConfig,
     inner_command: String,
 ) -> Result<(), Error> {
     config.image.obtain_image()?;
-
     let subshell_command = command::subshell_command(&config.init, inner_command);
     debug!("Running container with command '{}'", &subshell_command);
-    interpret::run_container(&environ, &floki_root, &config, &subshell_command)
-}
-
-/// Search all ancestors of the current directory for a floki.yaml file name.
-fn find_floki_yaml(current_directory: &path::Path) -> Result<path::PathBuf, Error> {
-    current_directory
-        .ancestors()
-        .map(|a| a.join("floki.yaml"))
-        .find(|f| f.is_file())
-        .ok_or(errors::FlokiError::ProblemFindingConfigYaml {}.into())
-}
-
-/// Take a file path, and return a tuple consisting of it's directory and the file path
-fn locate_file_in_system(path: path::PathBuf) -> Result<(path::PathBuf, path::PathBuf), Error> {
-    let dir = path
-        .parent()
-        .ok_or_else(|| errors::FlokiInternalError::InternalAssertionFailed {
-            description: format!("config_file '{:?}' does not have a parent", &path),
-        })?
-        .to_path_buf();
-    Ok((dir, path))
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use std::fs;
-    use tempdir;
-
-    fn touch_file(path: &path::Path) -> Result<(), Error> {
-        fs::create_dir_all(
-            path.parent()
-                .ok_or(format_err!("Unable to take parent of path"))?,
-        )?;
-        fs::OpenOptions::new().create(true).write(true).open(path)?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_floki_yaml_current_dir() -> Result<(), Error> {
-        let tmp_dir = tempdir::TempDir::new("")?;
-        let floki_yaml_path = tmp_dir.path().join("floki.yaml");
-        touch_file(&floki_yaml_path)?;
-        assert_eq!(find_floki_yaml(&tmp_dir.path())?, floki_yaml_path);
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_floki_yaml_ancestor() -> Result<(), Error> {
-        let tmp_dir = tempdir::TempDir::new("")?;
-        let floki_yaml_path = tmp_dir.path().join("floki.yaml");
-        touch_file(&floki_yaml_path)?;
-        assert_eq!(
-            find_floki_yaml(&tmp_dir.path().join("dir/subdir"))?,
-            floki_yaml_path
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_floki_yaml_sibling() -> Result<(), Error> {
-        let tmp_dir = tempdir::TempDir::new("")?;
-        let floki_yaml_path = tmp_dir.path().join("src/floki.yaml");
-        touch_file(&floki_yaml_path)?;
-        assert!(find_floki_yaml(&tmp_dir.path().join("include")).is_err());
-        Ok(())
-    }
+    interpret::run_container(&environ, &config, &subshell_command)
 }
